backend/requirements.txt
wheel
black
isort
pytest
pytest-django

nh3~=0.2
celery~=5.4
Django~=5.1
django-extensions~=3.2
django-filter~=24.3
django-phonenumber-field~=8.0
djangorestframework~=3.15
fuzzywuzzy~=0.18
google-auth~=2.35
Markdown~=3.7
phonenumberslite~=8.13
psycopg2-binary~=2.9
python-dateutil~=2.9
python-Levenshtein~=0.26
rich~=13.9
requests~=2.32
Werkzeug~=3.0
